Support for Path Variables & Content-Type Agnosticism
Path variables are now supported. The path templates are replaced by a
templating engine in cl-strings, so this is a transitive dependency
for now. I will aim to remove this at some point.

It no longer made sense to check which media types were supported
since we allow the consumer of the generated client to pass in
functions which will encode any media type. Any such checks have been
removed.

There was a minor bugfix to handle the case when a schema doesn't
provide any properties.